transfer fix

+++ Vortex2D/Engine/LinearSolver/GaussSeidel.cpp
@@ -11,7 +11,7 @@ namespace Vortex2D { namespace Fluid {
 
 
 GaussSeidel::GaussSeidel(const Renderer::Device& device, const glm::ivec2& size)
-    : mW(2.0f/(1.0f+std::sin(glm::pi<float>()/std::sqrt(size.x*size.y))))
+    : mW(2.0f/(1.0f+std::sinf(glm::pi<float>()/std::sqrtf((float)(size.x*size.y)))))
     , mGaussSeidel(device, size, "../Vortex2D/GaussSeidel.comp.spv",
                    {vk::DescriptorType::eStorageBuffer,
                     vk::DescriptorType::eStorageBuffer,
